Fix validation loader and classification dtype issues
- Set drop_last=False for val_loader to prevent empty batches
  when validation set size < batch_size
- Convert predictions and targets to long dtype for binary
  classification metrics (Accuracy, Precision, Recall)

+++ train.py
@@ -243,7 +243,8 @@ def train_dgl(config: Union[TrainingConfig, Dict[str, Any]], model: nn.Module = 
         # Classification metrics
         def thresholded_output_transform(output):
             y_pred, y = output
-            y_pred = (y_pred > 0.5).float()
+            y_pred = (y_pred > 0.5).long()  # Convert to long for binary classification metrics
+            y = y.long()  # Ensure targets are also long
             return y_pred, y
 
         metrics = {
